requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2  # Required for HTML template rendering (test UI)

# Database
motor==3.3.2
beanie==1.23.6
pymongo==4.6.3

# Authentication
python-jose[cryptography]==3.3.0  # Legacy JWT support
PyJWT==2.8.0  # Modern JWT validation (OAuth 2.0)
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.0

# Logging (Production-grade structured logging)
structlog==24.1.0
PyYAML==6.0.1
colorama==0.4.6  # For colored console output in development

# Security & Rate Limiting
slowapi==0.1.9  # Rate limiting for FastAPI
bleach==6.1.0  # HTML sanitization for XSS prevention

# Caching (optional)
redis==5.0.1  # Redis client for caching

# Monitoring & Metrics
prometheus-fastapi-instrumentator==6.1.0  # Prometheus metrics
psutil==5.9.8  # System and process utilities for dashboard metrics

# Testing (development only)
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1  # Async HTTP client for Auth API integration (native asyncio, no anyio DNS issues)
pytest-cov==4.1.0  # Code coverage
